\name{arab}
\docType{data}
\alias{arab}
\title{Arabidopsis thaliana data set from Atwell et al. 2010 \emph{Nature}}
\description{
The \sQuote{arab} data contains 107 phenotypes of Arabidopsis and genotypes from a 250K SNP chip as GenABEL gwaa.data format. 
}
\format{ The \sQuote{arab} data set contains a \code{gwaa.data} object. See package GenABEL.
}
\source{Atwell, S., Y. S. Huang, B. J. Vilhjalmsson, G. Willems, M. Horton, et al., 2010. Genome-wide association study of 107 phenotypes in Arabidopsis thaliana inbred lines. \emph{Nature} 465: 627-631.}
\references{Atwell, S., Y. S. Huang, B. J. Vilhjalmsson, G. Willems, M. Horton, et al., 2010. Genome-wide association study of 107 phenotypes in Arabidopsis thaliana inbred lines. \emph{Nature} 465: 627-631.}
\keyword{datasets}