% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p.exact.binary}
\alias{p.exact.binary}
\title{Exact p-values for genome-wide association analysis of case-control data in inbred populations}
\usage{
p.exact.binary(pheno, gwaa.object, or = NULL, or.maf = NULL,
  low.maf = 0.05, high.ld = 0.9, method = "logOR", type = "two-sided",
  con.table = NULL)
}
\arguments{
\item{pheno}{A string that gives the binary phenotype name in \code{gwaa.object} or
a vector that gives phenotypic values match the order in \code{gwaa.object}.}

\item{gwaa.object}{An object of \code{\link{gwaa.data-class}} to be analyzed.}

\item{or}{An (optional) vector gives the odds ratios to be tested, and \code{or.maf} must
also be given if not testing for the whole genome.}

\item{or.maf}{An (optional) vector gives the minor allele frequencies in accordance with
\code{or}, must be used together with \code{or}. If \code{or.maf} or \code{or}
is missing, exact p-values will be calculated for the whole genome.}

\item{low.maf}{A numeric value that gives the cut-off of the lowest minor allele frequency
allowed in the analysis.}

\item{high.ld}{A numeric value that gives the cut-off of the highest linkage disequilibrium R-square
allowed in the analysis, i.e. LD pruning.}

\item{method}{A string tells the method used, currently only "logOR" is available.}

\item{type}{A string tells the statistical test type, can be \code{'one-sided'} or \code{'two-sided'}.}

\item{con.table}{Genome-wide contingency tables. If \code{NULL}, to be calculated based on data.
If a string tells the file name, load from the file or to be calculated and saved into the file. If an R
matrix, regard as genome-wide contingency tables (see the saved R object for the format).}
}
\value{
The function returns a data frame of exact p-values (\code{$p.exact}) and corresponding odds
ratios \code{$or} and MAFs \code{$MAF}.
}
\description{
The function imports GenABEL (gwaa.data class) data format
and calculates the exact dataset-specific p-values of a case-control
phenotype for each variant or a given odds ratio and allele frequency.
}
\note{
None.
}
\examples{
\dontrun{
## loading example gwaa.data of data from Atwell et al. (2010) Nature
data(arab)

## running a regular GWA analysis for AvrRPM1
cc1 <- ccfast('X33_.i.avrRpm1..i.', data = arab)

## check the top finding using the exact p-value
top <- which.min(cc1[,'P1df'])
or <- cc1[top,'effB']
f <- summary(arab[,top])$Q.2
maf <- min(f, 1 - f)
exact <- p.exact.binary(pheno = 'X33_.i.avrRpm1..i.', gwaa.object = arab, or = or, or.maf = maf, con.table = 'tab.AvrRPM1.RData')
}
}
\author{
Xia Shen
}
\references{
Xia Shen (2015). Beyond permutation test: calculating exact dataset-specific p-values
for genome-wide association studies in inbred populations. \emph{Submitted}.
}
\seealso{
\code{\link{ccfast}}, \code{\link{glm}}
}
\keyword{discovery}
\keyword{exact}
\keyword{false}
\keyword{genome-wide}
\keyword{p-value,}
\keyword{p.exact,}
\keyword{rate}

