% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p.exact.gaussian}
\alias{p.exact.gaussian}
\title{Exact p-values for genome-wide association analysis in inbred populations}
\usage{
p.exact.gaussian(gwaa.object = NULL, n = NULL, maf = NULL, beta,
  beta.maf = NULL, low.maf = 0.05, type = "two-sided")
}
\arguments{
\item{gwaa.object}{An (optional) object of \code{\link{gwaa.data-class}}.}

\item{n}{An (optional) integer gives the sample size, only used when \code{gwaa.object = NULL}.}

\item{maf}{An (optional) vector gives minor allele frequencies across the genome,
only used when \code{gwaa.object = NULL}.}

\item{beta}{An vector gives the effect sizes to be tested, and \code{beta.maf} must
also be given if not testing for the whole genome.}

\item{beta.maf}{An (optional) vector gives the minor allele frequencies in accordance with
\code{beta}, must be used together with \code{beta}. If \code{beta.maf}
is missing, exact p-values will be calculated for the whole genome.}

\item{low.maf}{A numeric value that givens the cut-off of the lowest minor allele frequency
allowed in the analysis.}

\item{type}{A string tells the statistical test type, can be \code{'one-sided'} or \code{'two-sided'}.}
}
\value{
The function returns a data frame of exact p-values (\code{$p.exact}) and corresponding effect
sizes \code{$beta} and MAFs \code{$MAF}.
}
\description{
The function imports GenABEL (gwaa.data class) data format
and calculates the exact dataset-specific p-values for each variant
or a given effect size and allele frequency.
}
\note{
None.
}
\examples{
\dontrun{
## loading example gwaa.data in GenABEL
data(ge03d2ex.clean)

## running a regular GWA analysis for height
qt1 <- qtscore(height, data = ge03d2ex.clean)

## running the multivariate GWAS again
exact <- p.exact.gaussian(ge03d2ex.clean, beta = qt1[,'effB'])
}
}
\author{
Xia Shen
}
\references{
Xia Shen (2015). Beyond permutation test: calculating exact dataset-specific p-values
for genome-wide association studies in inbred populations. \emph{Submitted}.
}
\seealso{
\code{\link{qtscore}}, \code{\link{t.test}}
}
\keyword{discovery}
\keyword{exact}
\keyword{false}
\keyword{genome-wide}
\keyword{p-value,}
\keyword{p.exact,}
\keyword{rate}

